+++ rspace_client/client_base.py
@@ -99,7 +99,9 @@ class ClientBase:
         """
         numeric_id = self._get_numeric_record_id(resource_id)
         return self.retrieve_api_results(
-            "/{}/{}".format(path, numeric_id), content_type=None, request_type="DELETE",
+            "/{}/{}".format(path, numeric_id),
+            content_type=None,
+            request_type="DELETE",
         )
 
     def retrieve_api_results(
@@ -205,7 +207,9 @@ class ClientBase:
         print(msg, file=sys.stderr)
 
     def _stream(
-        self, endpoint: str, pagination: Pagination = Pagination(),
+        self,
+        endpoint: str,
+        pagination: Pagination = Pagination(),
     ):
         """
         Yields items, making paginated requests to the server as each page