+++ rspace_client/inv/inv.py
@@ -980,7 +980,7 @@ class ImageContainerPost(ContainerPost):
         image_file_path : str
             A full file path to the image to use.
         locations : Optional[Sequence], optional
-            An optional list of (x,y) coordinate tuples of marked locations within 
+            An optional list of (x,y) coordinate tuples of marked locations within
             the image container.
         tags : Optional[str], optional
             Comma separated tags
@@ -1206,7 +1206,9 @@ class InventoryClient(ClientBase):
         if sample_filter is not None:
             pagination.data.update(sample_filter.data)
         return self.retrieve_api_results(
-            f"/{endpoint}", request_type="GET", params=pagination.data,
+            f"/{endpoint}",
+            request_type="GET",
+            params=pagination.data,
         )
 
     def stream_samples(
@@ -1436,7 +1438,8 @@ class InventoryClient(ClientBase):
         id_to_copy = Id(item_to_duplicate)
         endpoint = id_to_copy.get_api_endpoint()
         rc = self.retrieve_api_results(
-            f"/{endpoint}/{id_to_copy.as_id()}/actions/duplicate", request_type="POST",
+            f"/{endpoint}/{id_to_copy.as_id()}/actions/duplicate",
+            request_type="POST",
         )
         if new_name is not None:
             rc = self.rename(rc, new_name)
@@ -1477,7 +1480,9 @@ class InventoryClient(ClientBase):
             raise ValueError("Supplied id is not a subsamples")
         data = {"content": note}
         return self.retrieve_api_results(
-            f"/subSamples/{ss_id.as_id()}/notes", request_type="POST", params=data,
+            f"/subSamples/{ss_id.as_id()}/notes",
+            request_type="POST",
+            params=data,
         )
 
     def get_workbenches(self) -> Sequence[dict]:
@@ -1703,7 +1708,11 @@ class InventoryClient(ClientBase):
             raise ValueError("Target must be a container")
         return id_target
 
-    def add_items_to_list_container(self, target_container_id, *item_ids: str,) -> list:
+    def add_items_to_list_container(
+        self,
+        target_container_id,
+        *item_ids: str,
+    ) -> list:
         """
         Adds 1 or more items to a list container
 
@@ -2093,7 +2102,8 @@ class InventoryClient(ClientBase):
         """
         id_to_restore = Id(sample_template_id)
         return self.retrieve_api_results(
-            f"/sampleTemplates/{id_to_restore.as_id()}/restore", request_type="PUT",
+            f"/sampleTemplates/{id_to_restore.as_id()}/restore",
+            request_type="PUT",
         )
 
     def transfer_sample_template_owner(